all: move NodeSet from light to trie
